requirements.txt
pydantic
PyYAML
pytest
pytest-playwright
allure-pytest
